+++ source/isaaclab/isaaclab/assets/asset_base_cfg.py
@@ -34,9 +34,9 @@ class AssetBaseCfg:
         # root position
         pos: tuple[float, float, float] = (0.0, 0.0, 0.0)
         """Position of the root in simulation world frame. Defaults to (0.0, 0.0, 0.0)."""
-        rot: tuple[float, float, float, float] = (1.0, 0.0, 0.0, 0.0)
-        """Quaternion rotation (w, x, y, z) of the root in simulation world frame.
-        Defaults to (1.0, 0.0, 0.0, 0.0).
+        rot: tuple[float, float, float, float] = (0.0, 0.0, 0.0, 1.0)
+        """Quaternion rotation (x, y, z, w) of the root in simulation world frame.
+        Defaults to (0.0, 0.0, 0.0, 1.0).
         """
 
     class_type: type[AssetBase] = None