+++ source/isaaclab/isaaclab/sim/spawners/wrappers/wrappers.py
@@ -37,7 +37,7 @@ def spawn_multi_asset(
         prim_path: The prim path to spawn the assets.
         cfg: The configuration for spawning the assets.
         translation: The translation of the spawned assets. Default is None.
-        orientation: The orientation of the spawned assets in (w, x, y, z) order. Default is None.
+        orientation: The orientation of the spawned assets in (x, y, z, w) order. Default is None.
         clone_in_fabric: Whether to clone in fabric. Default is False.
         replicate_physics: Whether to replicate physics. Default is False.
 
@@ -145,7 +145,7 @@ def spawn_multi_usd_file(
         prim_path: The prim path to spawn the assets.
         cfg: The configuration for spawning the assets.
         translation: The translation of the spawned assets. Default is None.
-        orientation: The orientation of the spawned assets in (w, x, y, z) order. Default is None.
+        orientation: The orientation of the spawned assets in (x, y, z, w) order. Default is None.
         clone_in_fabric: Whether to clone in fabric. Default is False.
         replicate_physics: Whether to replicate physics. Default is False.