+++ source/isaaclab_assets/isaaclab_assets/robots/shadow_hand.py
@@ -46,7 +46,7 @@ SHADOW_HAND_CFG = ArticulationCfg(
     ),
     init_state=ArticulationCfg.InitialStateCfg(
         pos=(0.0, 0.0, 0.5),
-        rot=(0.0, 0.0, -0.7071, 0.7071),
+        rot=(0.0, -0.7071, 0.7071, 0.0),
         joint_pos={".*": 0.0},
     ),
     actuators={