+++ source/isaaclab_tasks/isaaclab_tasks/direct/automate/disassembly_env_cfg.py
@@ -162,7 +162,7 @@ class DisassemblyEnvCfg(DirectRLEnvCfg):
                 "panda_finger_joint2": 0.04,
             },
             pos=(0.0, 0.0, 0.0),
-            rot=(1.0, 0.0, 0.0, 0.0),
+            rot=(0.0, 0.0, 0.0, 1.0),
         ),
         actuators={
             "panda_arm1": ImplicitActuatorCfg(