+++ source/isaaclab_assets/docs/CHANGELOG.rst
@@ -1,6 +1,15 @@
 Changelog
 ---------
 
+0.3.0 (2026-01-30)
+~~~~~~~~~~~~~~~~~~~
+
+Changed
+^^^^^^^
+
+* Changed the quaternion ordering to match warp, PhysX, and Newton native XYZW quaternion ordering.
+
+
 0.2.4 (2025-11-26)
 ~~~~~~~~~~~~~~~~~~